+++ CHANGELOG.md
@@ -13,6 +13,7 @@ incremented for features.
 
 ### Breaking
 
+* lang: Enforce that the payer for an init-ed account be marked `mut` ([#1271](https://github.com/project-serum/anchor/pull/1271)).
 * ts: `BorshAccountsCoder.accountDiscriminator` method has been replaced with `BorshAccountHeader.discriminator` ([#]()).
 * lang, ts: 8 byte account discriminator has been replaced with a versioned account header ([#]()).